+++ src/bluesearch/entrypoint/database/topic_extract.py
@@ -121,11 +121,16 @@ def run(
     Parameter description and potential defaults are documented inside of the
     `get_parser` function.
     """
+    import datetime
+
     from defusedxml import ElementTree
 
+    import bluesearch
     from bluesearch.database.topic import (
         extract_article_topics_for_pubmed_article,
+        extract_article_topics_from_medrxiv_article,
         extract_journal_topics_for_pubmed_article,
+        get_topics_for_arxiv_articles,
         get_topics_for_pmc_article,
     )
     from bluesearch.database.topic_info import TopicInfo
@@ -172,6 +177,49 @@ def run(
                 if journal_topics:
                     topic_info.add_journal_topics("MeSH", journal_topics)
                 all_results.append(topic_info.json())
+    elif source == "arxiv":
+        all_results = [
+            {
+                "source": "arxiv",
+                "path": str(path.resolve()),
+                "topics": {
+                    "article": {
+                        "arXiv": article_topics,
+                    },
+                },
+                "metadata": {
+                    "created-date": datetime.datetime.now().strftime(
+                        "%Y-%m-%d %H:%M:%S"
+                    ),
+                    "bbs-version": bluesearch.version.__version__,
+                },
+            }
+            for path, article_topics in get_topics_for_arxiv_articles(inputs).items()
+        ]
+
+    elif source in {"biorxiv", "medrxiv"}:
+        for path in inputs:
+            logger.info(f"Processing {path}")
+            topic, journal = extract_article_topics_from_medrxiv_article(path)
+            all_results.append(
+                {
+                    "source": journal,
+                    "path": str(path.resolve()),
+                    "topics": {
+                        "article": {
+                            "Subject Area": topic,
+                        },
+                    },
+                    "metadata": {
+                        "created-date": datetime.datetime.now().strftime(
+                            "%Y-%m-%d %H:%M:%S"
+                        ),
+                        "bbs-version": bluesearch.version.__version__,
+                    },
+                }
+            )
+
+        pass
     else:
         logger.error(f"The source type {source!r} is not implemented yet")
         return 1